Mojom compiler frontend: Change the type of the |offset| field to uint32.
The |offset| field within the |StructField| struct in mojom_types.mojom
should be a uint32 but it was previously an int32.

R=vardhan@google.com

Review URL: https://codereview.chromium.org/2005343003 .

+++ mojom/mojom_tool/serialization/serialization.go
@@ -322,9 +322,7 @@ func translateStructField(f *mojom.StructField) (field mojom_types.StructField) 
 		field.DefaultValue = translateDefaultFieldValue(f.DefaultValue)
 	}
 	if emitComputedPackingData {
-		// TODO(rudominer) Check the allowed size of offsets. The type used in
-		// mojom_types.mojom might need to be changed.
-		field.Offset = int32(f.Offset())
+		field.Offset = f.Offset()
 		field.Bit = int8(f.Bit())
 		field.MinVersion = f.MinVersion()
 	}
